Dec_06_2013 Updates

+++ assets/js/stream.js
@@ -32,13 +32,9 @@ function submit_note_data(elt) {
     var timestamp=getTimeStamp();
     var divtext=document.getElementById("note_creator_div");
     
-    
     var apiurl = "&uid="+enco(uid)+"&content_type=note&lat="+enco(lat)+"&long="+enco(long1)+"&timestamp="+timestamp;
     //console.log(apiurl);
     var postData = {note_text:enco(nl2br(note_text))};
-    //console.log(postData);
-    //&note_text=God%20gives%20as%20much%20as%20we%20can%20satisfy%20for%20life
-    //&lat=77&long=33&timestamp=2013-02-01%2022:11:00
     //call note api
     $.post(site_url+"api/write/?action_object=single_content"+apiurl,postData,function(rdata) {
         if(rdata.status == "success") {
@@ -59,7 +55,6 @@ function fail(rdata){
     console.log(rdata.responseText);
 }
 function submit_reminder_data(elt) {
-    //$(elt).
     var reminder_name = $("#reminder_title").val();
     var reminder_date = $("#reminder_date").val();
     var reminder_time = $("#reminder_time").val();
@@ -70,9 +65,7 @@ function submit_reminder_data(elt) {
     if(uid == '') {alert("Please Sign-In."); location=site_url+"?Please Sign-In."; return false; }
     
     var remind_time=reminder_date+" "+reminder_time;
-    //store this data 
-    //along with lat long info
-    //store this data 
+    //store this data along with lat long info
     var lat='77';
     var long1='23';
     var timestamp=getTimeStamp();
@@ -83,10 +76,8 @@ function submit_reminder_data(elt) {
         +"&timestamp="+timestamp;
     //console.log(apiurl);
     var postData = {remind_time:remind_time,reminder_name:enco(reminder_name)};
-    console.log(postData);
+    //console.log(postData);
     
-    //&content_type=reminder&remind_time=2013-11-07%2001:23:22
-    //&reminder_name=God%20textfor%20develpment%20ashhd&lat=77&long=33&timestamp=2013-11-06%2022:11:00
     $.post(site_url+"api/write/?action_object=single_content"+apiurl,postData,function(rdata) {
         if(rdata.status == "success") {
             //alert("Reminder Added.");
@@ -114,22 +105,15 @@ function submit_expense_data(elt) {
     var uid = $("#uid").val();
     if(uid == '') {alert("Please Sign-In."); location=site_url+"?Please Sign-In."; return false; }
     
-    //store this data 
-    //along with lat long info
-    
-    //store this data 
+    //store this data, along with lat long info
     var lat='77';
     var long1='23';
     var timestamp=getTimeStamp();
     
     var apiurl = "&uid="+enco(uid)+"&content_type=expense&lat="+enco(lat)+"&long="+enco(long1)+"&timestamp="+timestamp;
-    //console.log(apiurl);
+    
     var postData = {title:enco(expense_title),desc:enco(expense_title),amount:(expense_amount)};
-    //console.log(postData);
-    //return false;
-    //api/write/?action_object=single_content&uid=54694568990687&content_type=expense
-    //&title=God%20gives&desc=God%20textfor%20develpment%20ashhd&amount=1000&lat=77
-    //&long=33&timestamp=2013-02-01%2022:11:00
+    
     $.post(site_url+"api/write/?action_object=single_content"+apiurl,postData,function(rdata) {
         if(rdata.status == "success") {
             //alert("Expense Info Added.");
@@ -143,7 +127,7 @@ function submit_expense_data(elt) {
     
     return false;
 }
-
+var default_entries=35;
 function loadStreamData() {
     var uid = $("#uid").val();
     
@@ -173,8 +157,8 @@ function loadStreamData() {
                                                         var expense_name=expense.expense_title;
                                                         var expense_amount = expense.expense_amount;
                                                         var content_id = expense.content_id;
-                                                        content_type='expense';
-                                                        note_options_req='yes';
+                                                        var content_type='expense';
+                                                        var note_options_req='yes';
 
                                                         str_exp+="<li class='single_expense'>\n\
                                                                     <span class=''>"+expense_name+"</span>\n\
@@ -198,7 +182,7 @@ function loadStreamData() {
                         },"json");
                 }
             }
-            // ====================
+            // REMINDERS INFORMATION
             if(rdata.reminders != null) {
                 var total_reminders = (rdata.reminders).length;
                 $("#ttl_reminders").html(total_reminders);
@@ -257,44 +241,53 @@ function loadStreamData() {
                     //END REMINDER CODE
             }
             
-            
+            //NOTES INFORMATION
             if(content_target_src=='stream') {
-                    //NOTES CODE
                     var note_output='';
                     
                     if(rdata.notes != null) {
                         var max_notes_count = (rdata.notes).length;
+                        
                         if(max_notes_count==0)
                         {
                                 var max_notes_count=1;
-                                var note_text='';
+                                var note_output='Add Note';
                                 var note_image='';
                         }
-                        $.each(rdata.notes,function(i,note) {
-                                var content_id= note.content_id;
-                                var note_id= note.note_id;
-                                var note_text = note.note_text;
-                                var note_image='';
-                                var note_options_req='yes';
-                                note_output += "<li class='pin single_note_card'>\n\
-                                                    <img src='"+note_image+"' />\n\
-                                                    <div class='editable_block' id='editor_"+content_id+"'>"+nl2br(note_text)+"</div>";
-                                                    
-                                                    if(note_options_req=='yes') {
-                                                           note_output += "<div>\n\
-                                                                       <ul class='note-options'>\n\
-                                                                                <li class='note-options-single delete_icon fl_le' id=\"deleteBtn_"+content_id+"\"><a href='javascript:void(0)' onclick=\"delete_this(this,'"+content_id+"','note')\">&nbsp</a></li>\n\
-                                                                                <li class='note-options-single edit_icon fl_le' id=\"editorBtn_"+content_id+"\" onclick=\"edit_this('"+content_id+"')\">&nbsp</li>\n\
-                                                                                \n\
-                                                                                <li class='note-options-single save_tick_icon fl_le hide' id='editor_submit_"+content_id+"' onclick=\"save_edit_text(this,'"+content_id+"','note')\"></li>\n\
-                                                                                <li class='note-options-single cancel_icon fl_le hide' id='editor_cancel_"+content_id+"'></li>\n\
-                                                                        </ul>\n\
-                                                               </div>";
-                                                   }
+                        else {
+                            
+                            //loadNotesInfo(rdata.notes,max_notes_count);
+                            
+                                $.each(rdata.notes,function(i,note) {
+                                    
+                                        if(i < default_entries) {
+                                            var content_id= note.content_id;
+                                            var note_id= note.note_id;
+                                            var note_text = note.note_text;
+                                            var note_image='';
+                                            var note_options_req='yes';
+                                            note_output += "<li class='pin single_note_card'>\n\
+                                                                <img src='"+note_image+"' />\n\
+                                                                <div class='' id='editor_"+content_id+"'>"+nl2br(note_text)+"</div>";
 
-                                note_output +="</li>";
+                                                                if(note_options_req=='yes') {
+                                                                       note_output += "<div>\n\
+                                                                                   <ul class='note-options'>\n\
+                                                                                            <li class='note-options-single delete_icon fl_le' id=\"deleteBtn_"+content_id+"\"><a href='javascript:void(0)' onclick=\"delete_this(this,'"+content_id+"','note')\">&nbsp</a></li>\n\
+                                                                                            <li class='note-options-single edit_icon fl_le' id=\"editorBtn_"+content_id+"\" onclick=\"edit_this('"+content_id+"')\">&nbsp</li>\n\
+                                                                                            \n\
+                                                                                            <li class='note-options-single save_tick_icon fl_le hide' id='editor_submit_"+content_id+"' onclick=\"save_edit_text(this,'"+content_id+"','note')\"></li>\n\
+                                                                                            <li class='note-options-single cancel_icon fl_le hide' id='editor_cancel_"+content_id+"' onclick=\"cancel_edit_text('"+content_id+"');\"></li>\n\
+                                                                                    </ul>\n\
+                                                                           </div>";
+                                                               }
 
-                        });
+                                            note_output +="</li>";
+                                        }
+
+                                });
+                        }
+                        $("#max_notes_count").val(max_notes_count);
                     }
                     $(".all_note_list_box").html(note_output);
             }
@@ -323,7 +316,7 @@ function save_edit_text(elt,content_id,cont_type) {
     var elementSubmitBtn = $('#editor_submit_'+content_id);
     var elementCancelBtn = $('#editor_cancel_'+content_id);
     
-    element.contentEditable = 'false';
+    element.contentEditable = 'false';removeClass("editable_block edit_block"+content_id);
     elementSubmitBtn.addClass("hide");
     elementCancelBtn.addClass("hide");
     editorBtn.removeClass("hide");
@@ -343,13 +336,27 @@ function save_edit_text(elt,content_id,cont_type) {
     //console.log(postData);
     $.post(site_url+"api/modify/?action_object=single_content"+apiurl,postData,function(rdata) {
         if(rdata.status == "success") {
-            //alert("Reminder Deleted.");
             loadStreamData();
         }
         else {
             console.log("\n"+rdata.response);
         }
     },"json").fail(fail);
+}
+
+function cancel_edit_text(content_id) {
+    var editorBtn = $('#editorBtn_'+content_id);
+    var deleteBtn = $('#deleteBtn_'+content_id);
+    var element = document.getElementById('editor_'+content_id);
+    var elementSubmitBtn = $('#editor_submit_'+content_id);
+    var elementCancelBtn = $('#editor_cancel_'+content_id);
+    //editorBtn.addEventListener('click', function(e) {});
+     
+     element.contentEditable = 'false';removeClass("editable_block edit_block"+content_id);
+     editorBtn.removeClass("hide");
+     deleteBtn.removeClass("hide");
+     elementSubmitBtn.addClass("hide");
+     elementCancelBtn.addClass("hide");
 }
 
 function edit_this(content_id) {
@@ -359,19 +366,14 @@ function edit_this(content_id) {
     var element = document.getElementById('editor_'+content_id);
     var elementSubmitBtn = $('#editor_submit_'+content_id);
     var elementCancelBtn = $('#editor_cancel_'+content_id);
-    //editorBtn.addEventListener('click', function(e) {});
-     
+    
       if (element.isContentEditable) {
-        // Disable Editing
-//        element.contentEditable = 'false';
-//        elementSubmitBtn.addClass("hide");
-//        elementCancelBtn.addClass("hide");
-//        editorBtn.removeClass("hide");
-//        deleteBtn.removeClass("hide");
             
         // You could save any changes here.
       } else {
         element.contentEditable = 'true';
+        element.setAttribute("class","editable_block edit_block"+content_id);
+        
         elementSubmitBtn.removeClass("hide");
         elementCancelBtn.removeClass("hide");
         editorBtn.addClass("hide");
@@ -398,7 +400,105 @@ function delete_this(elt,content_id,cont_type) {
     },"json").fail(fail);
     
 }
+function removeClass(className) {
+    // convert the result to an Array object
+    var els = Array.prototype.slice.call(
+        document.getElementsByClassName(className)
+    );
+    for (var i = 0, l = els.length; i < l; i++) {
+        var el = els[i];
+        el.className = el.className.replace(
+            new RegExp('(^|\\s+)' + className + '(\\s+|$)', 'g'),
+            '$1'
+        );
+    }
+    
+}
+var scrollInAction = false;
+var limit_start = default_entries;
 
+function loadNotesInfo(limit_start,limit_end) {
+    var uid = $("#uid").val();
+    
+    var apiurl = "&uid="+enco(uid)+"&content_type=note&limit_start="+enco(limit_start)+"&limit_end="+enco(limit_end);
+    
+    $.post(site_url+"api/search/?action_object=list_content"+apiurl,{},function(rdata) {
+            var note_output='';
+            //console.log(rdata.notes);
+            $.each(rdata.notes,function(i,note) {
+                       
+                        var content_id= note.content_id;
+                        var note_id= note.note_id;
+                        var note_text = note.note_text;
+                        var note_image='';
+                        var note_options_req='yes';
+                        note_output += "<li class='pin single_note_card'>\n\
+                                            <img src='"+note_image+"' />\n\
+                                            <div class='' id='editor_"+content_id+"'>"+nl2br(note_text)+"</div>";
+
+                                            if(note_options_req=='yes') {
+                                                   note_output += "<div>\n\
+                                                               <ul class='note-options'>\n\
+                                                                        <li class='note-options-single delete_icon fl_le' id=\"deleteBtn_"+content_id+"\"><a href='javascript:void(0)' onclick=\"delete_this(this,'"+content_id+"','note')\">&nbsp</a></li>\n\
+                                                                        <li class='note-options-single edit_icon fl_le' id=\"editorBtn_"+content_id+"\" onclick=\"edit_this('"+content_id+"')\">&nbsp</li>\n\
+                                                                        \n\
+                                                                        <li class='note-options-single save_tick_icon fl_le hide' id='editor_submit_"+content_id+"' onclick=\"save_edit_text(this,'"+content_id+"','note')\"></li>\n\
+                                                                        <li class='note-options-single cancel_icon fl_le hide' id='editor_cancel_"+content_id+"' onclick=\"cancel_edit_text('"+content_id+"');\"></li>\n\
+                                                                </ul>\n\
+                                                       </div>";
+                                           }
+
+                        note_output +="</li>";
+                        
+                        
+                        scrollInAction=false;
+                        
+            });
+            
+            
+            $(".all_note_list_box").append(note_output);
+    },"json");
+    
+}
+
+  $(window).scroll(function(){
+        var max_notes_count= document.getElementById("max_notes_count").value;
+        
+        if(max_notes_count > 0) {
+                //console.log($(window).scrollTop() >= ($(document).height() - $(window).height() - 200));
+                   if  ($(window).scrollTop() >= ($(document).height() - $(window).height() - 200)){
+
+                         if (scrollInAction)
+                         {
+                             return false;
+                         }
+                         else
+                         {
+                                scrollInAction=true;
+                                //do stuff // load of more content here
+                                //console.log("Down");
+                                
+                                
+                                if(limit_start >= default_entries && limit_start <= max_notes_count) {
+                                        limit_start += 10;
+
+                                        if(max_notes_count >= limit_start) {
+
+                                            var limit_end=1;
+//                                            console.log(limit_start);
+                                            loadNotesInfo(limit_start,limit_end);
+                                             //scrollInAction=false; // TODO: PUT this scrollInAction=false after ajax loading is completed.
+                                        }
+                                        else { 
+                                            scrollInAction=false;
+                                        }
+                                }
+                         }
+                   }
+        }
+  });
+  
+     
 $(document).ready(function() {
     loadStreamData();
     //clearing form fields...
